requirements.txt
python-pptx
PyYAML
openai
pytest
